verdin-imx8mp: Properly configure eMMC devices, kernel and bootloader for insecure EFI boot.
For the Verdin iMX8MP board, the boot container and system image needs to be flashed on the internal eMMC.
The Toradex kernel does not come with proper dm-verity and squashfs configuration, so it needs additional fragments.
As the current state is based on latest-greatest OE layers, a breaking change in the output of veritysetup is reverted.